+++ helpers/helpers_xxx_global.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//05/06/26
+//15/06/26
 
 /* exported loadUserDefFile, addGlobValues, globFonts, globSettings, globNoSplitArtist */
 
@@ -338,7 +338,9 @@ const globSettings = {
 	bTooltip: true,
 	bLogToFile: false,
 	instanceManager: 'v1',
-	binariesPath: _foldPath(folders.binaries)
+	binariesPath: _foldPath(folders.binaries),
+	userAgent: 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/146.0.0.0 Safari/537.36',
+	curlImpersonate: 'curl_chrome146.bat'
 };
 
 // No-Split Artist: user replaceable with a presets file at folders.data