+++ CHANGELOG.md
@@ -78,6 +78,8 @@
 	* 'Playlist-Manager-SMP: switch enable panel'		-> { window: string[] }
 	* 'Playlist-Manager-SMP: enable panel'				-> { window: string[] }
 	* 'Playlist-Manager-SMP: disable panel'				-> { window: string[] }
+- Configuration: added new "userAgent" (Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/146.0.0.0 Safari/537.36) and "curlImpersonate" (curl_chrome146.bat) at global settings (globSettings.json) for arbitrary downloads using curl and/or JS-host methods.
+- Helpers: added curl-impersonate.exe v1.5.6. See [here](https://github.com/lexiforest/curl-impersonate).
 ### Changed
 - Quick-search: foobar2000 shortcuts are now skipped while quick-searching if JS-host supports it (for ex. to bypass play/pause when pressing space bar).
 - Quick-search: tooltip is now disabled while quick-searching.
@@ -93,6 +95,7 @@
 - Exporting Playlist: updated Fiio 'Export and convert' preset to split artist into artist initial subfolders (android doesn't handle a huge amount of folders at any path really well). Added new Namida preset.
 - Helpers: updated curl.exe to v8.20.0-slim. See [here](https://github.com/lordmulder/cURL-build-win32).
 - Helpers: updated 7za.exe to v26.01. See [here](https://www.7-zip.org/download.html).
+- Helpers: code cleanup.
 ### Removed
 - UI: loaded status icon being shown for UI-only playlists on lite mode despite such info being totally useless. See [here](https://hydrogenaudio.org/index.php/topic,120979.msg1082334.html#msg1082334).
 ### Fixed